Split the code into different modules for organization
