+++ constants.py
@@ -1,0 +1,14 @@
+"""
+Contains constant variables
+"""
+
+tile_size = 30
+fps = 60
+screen_width, screen_height = 500, 500
+
+# COLORS
+black = (0, 0, 0)
+dark_gray = (65, 64, 66)
+light_gray = (209, 211, 212)
+gray = (109, 110, 113)
+purple = (179, 136, 255)
